+++ tabsLayout.go
@@ -147,7 +147,7 @@ func (tabsLayout *tabsLayoutData) remove(tag string) {
 			return
 		}
 		if tabsLayout.created {
-			tabsLayout.session.runScript(fmt.Sprintf("activateTab(%v, %d);", tabsLayout.htmlID(), 0))
+			tabsLayout.session.runScript(fmt.Sprintf("activateTab('%v', %d);", tabsLayout.htmlID(), 0))
 			for _, listener := range tabsLayout.tabListener {
 				listener(tabsLayout, 0, oldCurrent)
 			}
@@ -224,7 +224,7 @@ func (tabsLayout *tabsLayoutData) set(tag string, value interface{}) bool {
 			return true
 		}
 		if tabsLayout.created {
-			tabsLayout.session.runScript(fmt.Sprintf("activateTab(%v, %d);", tabsLayout.htmlID(), current))
+			tabsLayout.session.runScript(fmt.Sprintf("activateTab('%v', %d);", tabsLayout.htmlID(), current))
 			for _, listener := range tabsLayout.tabListener {
 				listener(tabsLayout, current, oldCurrent)
 			}